refactor: refactor scaling module
